chore: prism-react-renderer version update
