+++ package.json
@@ -27,7 +27,7 @@
     "next-image-export-optimizer": "^1.12.3",
     "next-mdx-remote": "^4.4.1",
     "next-remote-watch": "^2.0.0",
-    "prism-react-renderer": "^1.3.5",
+    "prism-react-renderer": "^2.3.1",
     "prismjs": "^1.29.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
